multiplied unnits

+++ src/AliceInventory/Parser/ParserResponse.cs
@@ -2,7 +2,7 @@ namespace ConsoleApp {
     public class ParserResponse {
         public string ItemString { get; set; }
 
-        public int ItemCount { get; set; }
+        public double ItemCount { get; set; }
 
         public Unit Unit { get; set; }
     }
